refactor: change project's structure
