+++ src/components/App-header/index.js
@@ -1,0 +1,3 @@
+import AppHeader from './App-header';
+
+export default AppHeader;
