+++ src/components/Item-status-filter/index.js
@@ -1,0 +1,3 @@
+import ItemStatusFilter from './Item-status-filter';
+
+export default ItemStatusFilter;
