+++ src/components/Search-panel/index.js
@@ -1,0 +1,3 @@
+import SearchPanel from './Search-panel';
+
+export default SearchPanel;
