+++ src/components/Todo-list-item/index.js
@@ -1,0 +1,3 @@
+import TodoListItem from './Todo-list-item';
+
+export default TodoListItem;
