+++ src/components/Todo-list/index.js
@@ -1,0 +1,3 @@
+import TodoList from './Todo-list';
+
+export default TodoList;
