+++ src/components/Todo-stat/index.js
@@ -1,0 +1,3 @@
+import TodoStat from './Todo-stat';
+
+export default TodoStat;
